requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0

# Database Drivers
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
loguru>=0.7.0
openpyxl>=3.1.0

# Development (opcional)
jupyter>=1.0.0
ipykernel>=6.0.0
